* Include locally available chunks in the "chunks seen" statistics for completed/distributed copies. Patch by Josef Drexler.
* Added 'd.get_bitfield' command for retrieving the bitfield in hex format. Patch by Thomas Rosner. (PD)

* Fixed include headers for gcc-4.4.0. Patch by 'kloeri'.

* Switched a 'long long' with 'int64_t' in xmlrpc.cc to avoid incompatible types for some systems.


git-svn-id: svn://rakshasa.no/libtorrent/trunk/rtorrent@1092 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/rpc/xmlrpc.cc
@@ -98,7 +98,7 @@ xmlrpc_list_entry_to_value(xmlrpc_env* env, xmlrpc_value* src, int index) {
 
 #ifdef XMLRPC_HAVE_I8
   case XMLRPC_TYPE_I8:
-    long long v2;
+    int64_t v2;
     xmlrpc_read_i8(env, tmp, &v2);
     xmlrpc_DECREF(tmp);
     return v2;
@@ -244,7 +244,7 @@ xmlrpc_to_object(xmlrpc_env* env, xmlrpc_value* value, int callType, rpc::target
 
 #ifdef XMLRPC_HAVE_I8
   case XMLRPC_TYPE_I8:
-    long long v2;
+    int64_t v2;
     xmlrpc_read_i8(env, value, &v2);
       
     return torrent::Object((int64_t)v2);
